Update RS_Preprocessing.m
Changed settings in clean_raw so that I reject data again, instead of correcting them. Based on EEGLAB recommendations. Filter settings are now set to [0.75 1.25] so that it is actually 1 Hz, which it was not before.  Moved to import EOG channel coordinates, recommended having before ASR. Changed some comments. The next step is CSD. Might need a separate script for that.