Update RS_CSD_Transformation.m
Had to fix some things to get the loops to work. The nested for loop did not work correctly, but now it does. Clearing variables in between loops messed it up. However, I am getting checkset messages that the number of epochs do not match. Will deal with that tomorrow.